.project might be there - avoid exception if not

+++ postProcessIndoqaQuickstartBoot.js
@@ -31,7 +31,9 @@ const modifyPom = () => {
 
 const removeUnnecessaryFiles = () => {
   const projectFile = './indoqa-quickstart-boot/target/generated-sources/archetype/src/main/resources/archetype-resources/.project'
-  fs.unlinkSync(projectFile)
+  if (fs.existsSync(projectFile)) {
+    fs.unlinkSync(projectFile)
+  }
 
   console.log(`Removed ${projectFile}`)
 }
